pyproject.toml
[project]
name = "blog-diogenes-dev"
version = "0.1.0"
description = "A blog about programming, data engineering and even more random stuff"
readme = "./README.md"
requires-python = ">=3.12"

dependencies = [
  "markdown>=3.5",
  "mkdocs>=1.6.1,<2.0",
  "mkdocs-material>=9.7.1",
  "pathspec>=0.12.0",
  "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.uv]
native-tls = true
package = false

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[dependency-groups]
dev = ["rust-just>=1.46.0"]
